Make sure response message doesn't have any additional records
Otherwise, if the question has additional records, they would not
be in the response (because they're not parsed yet), but th count
would be different than zero, causing the response message to be
malformed (but still valid).

This fixes the tests failing on ubuntu, which seems to send
additional records on the query.