feature: Remove process management
I've decided to remove process management completely from zas. It
introduced a lot of complications, but the main reason is that it made
it very difficult to debug applications (with binding.pry, for
instance).

From now on, zas will only take care of routing. It has simplified the
whole system a lot and will allow cooler features to be developed.

Issues solved:
 - Closes #18 : Since now you start your own apps, there is no need to
 support custom commands. Even docker should work well integrated with
 other non docker or even remote apps.

 - Closes #17: No more process management.

 - Closes #15: No more spawning user shells

 - Closes #12: No more process management.

 - Closes #9: No more process management.

 - Closes #5: No more process management.

 - Closes #4: Removed foreman dependency.

 - Closes #2: It just fails for now, but no hanging.